deliverable2/requirements.txt
streamlit>=1.37.0
openai>=1.44.0
pyyaml>=6.0.1
